chore: move link to extensions
