+++ openmetadata-ui/src/main/resources/ui/src/components/BlockEditor/Extensions/index.ts
@@ -15,12 +15,12 @@ import Placeholder from '@tiptap/extension-placeholder';
 import TaskItem from '@tiptap/extension-task-item';
 import TaskList from '@tiptap/extension-task-list';
 import StarterKit from '@tiptap/starter-kit';
-import { LinkExtension } from '../link';
 import DiffView from './diff-view';
 import DragAndDrop from './drag-and-drop';
 import { Hashtag } from './hashtag';
 import { hashtagSuggestion } from './hashtag/hashtagSuggestion';
 import { Image } from './image/image';
+import { LinkExtension } from './link';
 import { Mention } from './mention';
 import { mentionSuggestion } from './mention/mentionSuggestions';
 import slashCommand from './slash-command';
